requirements.txt
asgiref==3.4.1
Django==3.2.14
django-mongodb-engine==0.6.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
djangotoolbox==1.8.0
dnspython==2.2.1
gunicorn==20.1.0
install==1.3.5
mongoengine==0.24.2
pkg_resources==0.0.0
PyJWT==2.4.0
pymongo==4.1.1
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
